doctor clinic data is done

+++ Doctoraak/View/CustomVC/Registration/Welcome/Main Login And Register/Clinic/CustomClinicDataView.swift
@@ -303,7 +303,7 @@ class CustomClinicDataView: CustomBaseView {
                 }
                 else {
                     
-                    clinicDataViewModel.waitingHours = texts
+                    clinicDataViewModel.waitingHours = texts.toInt()
                     floatingLabelTextField.errorMessage = ""
                 }
